abstract handlers
